Stepping motor driver using Pico & A4988
Using a Raspberry Pi Pico and an A4988 stepping motor driver, this program allows exact RPM (needed for a clock, for example) or a 10K pot to control the speed as desired.   In the A4988, the "Reset" and"Sleep" must be at Vcc of the chip (not the 12 volt to the motor).  My stepper draws 0.3 amps regardless of the speed, so be sure to use a large enough power supply.  https://bit.ly/3zimW3v gives the wiring for the driver.  The code above essentially gives the wiring for the Pico.